feat: add controller test instead of api route test
